requirements.txt
Django==5.2.6
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
mysqlclient==2.2.7
PyJWT==2.10.1
gunicorn==21.2.0
python-dotenv==1.0.0
